test(dampener): the on-disk baseline gets its own file, and the repository's own baseline is refreshed
tests/test_sense_dampener.py reached 855 lines, over the ratchet that fails a
file crossing 800. The division is the one the module already makes: the
arithmetic and its rendering stay, the artifact on disk moves to
tests/test_sense_dampener_baseline.py with the identity it carries and every
way reading it fails.

The committed baseline is regenerated on the merged tree, so it records sense
schema 3 (#335 bumped it) rather than a note about a schema that moved.
